+++ defaults/quickshell/Bar/Bar.qml
@@ -106,7 +106,6 @@ PanelWindow {
         anchors.leftMargin: root.theme.sectionPadding
         anchors.verticalCenter: parent.verticalCenter
 
-        ApplicationLauncherWidget { bar: root; theme: root.theme }
         SystemOverviewWidget { bar: root; theme: root.theme }
         BarDivider { theme: root.theme }
         WorkspacesWidget { bar: root; theme: root.theme }